pyproject.toml
[tool.poetry]
name = "pycasting"
version = "0.0.1a1"
description = "Python-based financial forecasting for startups."
authors = ["Austin Howard <austin@tangibleintelligence.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">3.8,<3.11"
pydantic = "^1.9.0"
uncertainties = "^3.1.6"
typer = "^0.4.0"
click = "8.0.4"
python-slugify = "^6.1.1"
python-dateutil = "^2.8.2"
clearcut = "^0.1.2"
frozendict = "^2.3.1"
pandas = "^1.4.2"
streamlit = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
